% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_create_safedir.R
\name{supp_create_safedir}
\alias{supp_create_safedir}
\title{supp_create_safedir}
\usage{
supp_create_safedir(safe.dir)
}
\arguments{
\item{safe.dir}{string (e.g. "~/home/..."). Specifies the path to a directory.}
}
\value{
The path to the newly created safe directory "/data__<Sys.Date>_<Sys.Time>".
}
\description{
Support method aiming to safeguard in process downloaded data from possible crashes of the program. Given a path to a
directory a "/data_<Sys.Date>_<Sys.Time>" folder is created with two sub-folders: "/content", storing each response's data in a
raw format (which can subsequently be parsed), and "/metadata", storing process summary metadata (e.g. next_token, total_counts, ect.).
}
\author{
: Victor Plesco
}
