% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_eval.ratelimit.R
\name{supp_eval.ratelimit}
\alias{supp_eval.ratelimit}
\title{supp_eval.ratelimit}
\usage{
supp_eval.ratelimit(request.remaining, request.reset)
}
\arguments{
\item{request.remaining}{integer. Number of requests left for the 15-minute window.}

\item{request.reset}{integer. Remaining window before the rate limit resets (in UTC epoch seconds).}
}
\description{
Manages the rate limits for the Twitter API v2 search endpoints.
}
\author{
Victor Plesco
}
