% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_eval.response.R
\name{supp_eval.response}
\alias{supp_eval.response}
\title{supp_eval.response}
\usage{
supp_eval.response(token, v2.endpoint, param_list)
}
\arguments{
\item{token}{object of class v2_create_token(). Bearer token associated with a user-created APP (requires a developer account).}

\item{v2.endpoint}{string \{"/2/tweets/search/all" or "/2/tweets/search/recent"\}. Twitter API v2 Search Endpoint.}

\item{param_list}{list (e.g. max_results = 10, ect.). List of a request's parameters.}
}
\value{
A list composed of a response's metadata and its parsed content (with attached errors if encountered).
}
\description{
Makes a request to a Twitter API v2 search endpoint and switches the return based on the encountered errors.
}
\author{
Victor Plesco
}
