% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_forward_request.R
\name{supp_forward_request}
\alias{supp_forward_request}
\title{supp_forward_request}
\usage{
supp_forward_request(token, v2.endpoint, param_list, safe.dir, sleep.work)
}
\arguments{
\item{sleep.work}{}
}
\value{

}
\description{

}
\author{
: Victor Plesco
}
