% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_get.conditions.R
\name{supp_get.conditions}
\alias{supp_get.conditions}
\title{supp_get.conditions}
\usage{
supp_get.conditions(v2.endpoint, product_track)
}
\arguments{
\item{v2.endpoint}{string \{"/2/tweets/search/all" or "/2/tweets/search/recent"\}. Twitter API v2 Search Endpoint.}

\item{product_track}{string \{"academic" or "standard"\}. Type of product track.}
}
\value{
A list of conditions.
}
\description{
Returns a request's conditions based on its endpoint and the type of product track used.
}
\author{
Victor Plesco
}
