% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_get.expansions.R
\name{supp_get.expansions}
\alias{supp_get.expansions}
\title{supp_get.expansions}
\usage{
supp_get.expansions(payload, expansions)
}
\arguments{
\item{payload}{string \{"tweet" or "user"\}. Type of Twitter payload.}

\item{expansions}{vector of strings (e.g. "expansion_\{1\}", ...). Expansions to be returned (accepts "all"). To request all but some 
expansions, append "-" in front of each expansion in the input vector (e.g. "-expansion_\{1\}", ...).}
}
\value{
A comma separated string of expansions (e.g. "author_id,geo.place_id").
}
\description{
Returns a customized set of expansions for a Twitter payload.
}
\seealso{
\url{https://developer.twitter.com/en/docs/twitter-api/expansions}
}
\author{
Victor Plesco
}
