% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_get.fields.R
\name{supp_get.fields}
\alias{supp_get.fields}
\title{supp_get.fields}
\usage{
supp_get.fields(object, fields)
}
\arguments{
\item{object}{string \{"tweet", "user", "media", "poll", "place"\}. Type of Twitter object.}

\item{fields}{vector of strings (e.g. "field_\{1\}", ...). Fields to be returned (accepts "all"). To request all but some fields, 
append "-" in front of each field in the input vector (e.g. "-field_\{1\}", ...).}
}
\value{
A comma separated string of fields (e.g. "author_id,geo.place_id").
}
\description{
Returns a customized set of fields of a Twitter object.
}
\seealso{
\url{https://developer.twitter.com/en/docs/twitter-api/fields}
}
\author{
Victor Plesco
}
