% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_get_endpoint_parameters.R
\name{supp_get_endpoint_parameters}
\alias{supp_get_endpoint_parameters}
\title{supp_get_endpoint_parameters}
\usage{
supp_get_endpoint_parameters(v2.endpoint, params)
}
\arguments{
\item{v2.endpoint}{string ("/2/___"). Specifies an endpoint for which a list of parameters is returned.}

\item{params}{vector of strings ("PARAM_1", ...). If specified, subsets the returned parameters (accepts "all").}
}
\value{
A list of parameters for the specified v2.endpoint having NULL as value.
}
\description{
Getter for the parameters of the endpoints available for request on the new Twitter API v2.
}
\author{
Victor Plesco
}
