% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_get_expansions.R
\name{supp_get_expansion_fields}
\alias{supp_get_expansion_fields}
\title{supp_get_expansion_fields}
\usage{
supp_get_expansion_fields(expansion, fields)
}
\arguments{
\item{expansion}{string \{"tweet" or "user"\}. Specifies an expansion for which a comma separated string of fields is returned.}

\item{fields}{vector of strings (e.g. "field_\{1\}", ...). Specifies a subset of fields to be returned (accepts "all"). To 
request all but some fields, append "-" in front of each field in the input vector (e.g. "-field_\{1\}", ...).}
}
\value{
A comma separated string of fields (e.g. "author_id,geo.place_id").
}
\description{
Getter for the fields of the expansions available for request on the Twitter API v2.
}
\author{
Victor Plesco
}
