% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_get_payload_expansions.R
\name{supp_get_payload_expansions}
\alias{supp_get_payload_expansions}
\title{supp_get_payload_expansions}
\usage{
supp_get_payload_expansions(payload, expansions)
}
\arguments{
\item{payload}{string \{"tweet" or "user"\}. Specifies a payload for which a comma separated string of expansions is returned.}

\item{expansions}{vector of strings (e.g. "expansion_\{1\}", ...). Specifies a subset of expansions to be returned (accepts "all"). 
To request all but some expansions, append "-" in front of each expansion in the input vector (e.g. "-expansion_\{1\}", ...).}
}
\value{
A comma separated string of expansions (e.g. "author_id,geo.place_id").
}
\description{
Getter for the expansions of a specified payload.
}
\seealso{
\url{https://developer.twitter.com/en/docs/twitter-api/expansions}
}
\author{
Victor Plesco
}
