% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_get_request_conditions.R
\name{supp_get_request_conditions}
\alias{supp_get_request_conditions}
\title{supp_get_request_conditions}
\usage{
supp_get_request_conditions(v2.endpoint, product_track)
}
\arguments{
\item{v2.endpoint}{string \{"/2/___"\}. Specifies an endpoint for which a list of query conditions is returned.}

\item{product_track}{string \{"academic" or "standard"\}. Specifies the type of developer account in possess to which specific
query conditions apply.}
}
\value{
A list of query conditions (e.g. max.max_results = 500, min.max_results = 10, ect.).
}
\description{
Getter for the query conditions of the combination of an endpoint and a product track.
}
\author{
: Victor Plesco
}
