% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_make.request.R
\name{supp_make.request}
\alias{supp_make.request}
\title{supp_make.request}
\usage{
supp_make.request(token, v2.endpoint, param_list, safe.dir = NULL)
}
\arguments{
\item{token}{object of class v2_create_token(). Bearer token associated with a user-created APP (requires a developer account).}

\item{v2.endpoint}{string \{"/2/tweets/search/all" or "/2/tweets/search/recent"\}. Twitter API v2 Search Endpoint.}

\item{param_list}{list (e.g. max_results = 10, ect.). List of a request's parameters.}

\item{safe.dir}{string (e.g. "~/home/..."). Default=NULL. Path to an existing directory (serving as a data backup). If NULL, considers the home directory.}
}
\value{
A response's parsed content (with attached errors if encountered).
}
\description{

}
\author{
Victor Plesco
}
