% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_make.safedir.R
\name{supp_make.safedir}
\alias{supp_make.safedir}
\title{supp_make.safedir}
\usage{
supp_make.safedir(safe.dir = NULL)
}
\arguments{
\item{safe.dir}{string (e.g. "~/home/..."). Default=NULL. Path to an existing directory (serving as a data backup). If NULL, considers the home directory.}
}
\value{
Path to the newly created directory.
}
\description{
Creates a "/racademic_data_<Sys.Date>_<Sys.Time>" directory with two sub-folders: "content" and "execution".
}
\author{
Victor Plesco
}
