% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_print_summary.R
\name{supp_print_summary}
\alias{supp_print_summary}
\title{supp_print_summary}
\usage{
supp_print_summary(process.data, twitter.chunk)
}
\arguments{
\item{process.data}{}

\item{twitter.chunk}{}
}
\value{

}
\description{

}
\author{
: Victor Plesco
}
