% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_validate_limitrate.R
\name{supp_validate_limitrate}
\alias{supp_validate_limitrate}
\title{supp_validate_limitrate}
\usage{
supp_validate_limitrate(round.time, limit.remaining, limit.reset)
}
\arguments{
\item{round.time}{POSIXlt (e.g. "2021-04-11 18:54:23 UTC"). Specifies the time of request.}

\item{limit.remaining}{integer. Specifies the number of requests left for the 15-minute window.}

\item{limit.reset}{integer (UTC epoch seconds). Specifies the remaining window before the rate limit resets.}
}
\description{
Support method for validating rate limits of an API request. Conceptually, the function must precede a second request
}
\author{
: Victor Plesco
}
