% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_validate_rate.R
\name{supp_validate_rate}
\alias{supp_validate_rate}
\title{supp_validate_rate}
\usage{
supp_validate_rate(round.time, limit.remaining, limit.reset)
}
\arguments{
\item{round.time}{POSIXlt (e.g. "2021-04-11 18:54:23 UTC"). Specifies the time of request.}

\item{limit.remaining}{integer. Specifies the number of requests left for the 15-minute window.}

\item{limit.reset}{integer. Specifies the remaining window before the rate limit resets, in UTC epoch seconds.}
}
\description{

}
\author{
: Victor Plesco
}
