% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_validate_response.R
\name{supp_validate_response}
\alias{supp_validate_response}
\title{supp_validate_response}
\usage{
supp_validate_response(token, v2.endpoint, param_list)
}
\arguments{
\item{v2.endpoint}{}

\item{param_list}{}
}
\value{

}
\description{

}
\author{
: Victor Plesco
}
