% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_create_bearer_token.R
\name{v2_create_bearer_token}
\alias{v2_create_bearer_token}
\title{v2_create_bearer_token}
\usage{
v2_create_bearer_token(v2.bearer_token)
}
\arguments{
\item{v2.bearer_token}{string. Specifies a bearer token.}
}
\value{
A bearer token object.
}
\description{
Creates a bearer token object.
}
\author{
: Victor Plesco
}
