% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_create_token.R
\name{v2_create_token}
\alias{v2_create_token}
\title{v2_create_token}
\usage{
v2_create_token(token)
}
\arguments{
\item{token}{string. Bearer token associated with a user-created APP (requires a developer account).}
}
\value{
An object of class racademic.
}
\description{
Creates an OAuth 2.0 Bearer authentication "token".
}
\author{
: Victor Plesco
}
