% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_search_fullarchive.R
\name{v2_search_fullarchive}
\alias{v2_search_fullarchive}
\title{v2_get_endpoint_parameters}
\usage{
v2_search_fullarchive(
  v2.token,
  product_track,
  query,
  start_time = NULL,
  end_time = NULL,
  max_results = 10,
  next_token = NULL,
  since_id = NULL,
  until_id = NULL,
  expansions = NULL,
  tweet.fields = NULL,
  user.fields = NULL,
  media.fields = NULL,
  place.fields = NULL,
  poll.fields = NULL
)
}
\arguments{
\item{v2.endpoint}{string ("/2/___"). . Specifies an endpoint for which a list of parameters is returned.}

\item{read.only}{boolean (TRUE, FALSE). Default=FALSE. Specifies the type of return of the function (see return for details).}

\item{custom.param}{vector of strings ("PARAM_1", ...). Default=NULL. Customizes the return, or print, by subsetting the parameters.}

\item{open.url}{boolean (TRUE, FALSE). Default=FALSE. If TRUE, opens the documentation page related to the v2.endpoint specified.}
}
\value{
A list of parameters of the specified v2.endpoint having NULL as value (if read.only is FALSE), or will print the parameters 
alongside with their description (if read.only is TRUE). If no v2.endpoint is specified, prints all the endpoints available for request.
}
\description{
Method of [v2_search_fullarchive] and [v2_search_recent()] functions that serves as a container for the parameters of 
the endpoints available for request on the new Twitter API v2.
}
\author{
Victor Plesco
}
