% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_create_safedir.R
\name{supp_create_safedir}
\alias{supp_create_safedir}
\title{supp_create_safedir}
\usage{
supp_create_safedir(safe.dir)
}
\arguments{
\item{safe.dir}{string (e.g. "~/home/..."). Specifies path to a directory.}
}
\value{
Path to the newly created directory "/data__<Sys.Date>_<Sys.Time>".
}
\description{
Support method for safeguarding in process downloaded data from possible crashes of the program. Given a path to an 
existing directory a "/data_<Sys.Date>_<Sys.Time>" folder is created with two sub-folders: "/content", aiming to store each response's 
parsed data in a .RDS file, and "/metadata", aiming to store process summary metadata (e.g. next_token, daily_counts, ect.).
}
\author{
: Victor Plesco
}
