% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_forward_request.R
\name{supp_forward_request}
\alias{supp_forward_request}
\title{supp_forward_request}
\usage{
supp_forward_request(token, v2.endpoint, param_list, safe.dir)
}
\arguments{
\item{token}{racademic's object. Specifies a bearer token associated with a user-created APP (requires a developer account).}

\item{v2.endpoint}{string \{"/2/___"\}. Specifies an endpoint to which a request is made.}

\item{param_list}{mix (e.g. max_results = 10, query = "btc -is:retweet", ect.). Specifies list of parameters to be used within request.}

\item{safe.dir}{string (e.g. "~/home/..."). Specifies path to a directory.}
}
\value{
If the request is successful returns a list with sub-lists "data", storing a tweet's data (each item is a tweet), and 
"includes", storing a tweet's expansions. If the request isn't successful appends, to the so far downloaded data, a sub-list "errors", 
storing the details of the errors in the request.
}
\description{

}
\author{
: Victor Plesco
}
