% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_get_conditions.R
\name{supp_get_conditions}
\alias{supp_get_conditions}
\title{supp_get_conditions}
\usage{
supp_get_conditions(v2.endpoint, product_track)
}
\arguments{
\item{product_track}{}
}
\value{

}
\description{

}
\author{
: Victor Plesco
}
