% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_get_object_fields.R
\name{supp_get_object_fields}
\alias{supp_get_object_fields}
\title{supp_get_object_fields}
\usage{
supp_get_object_fields(v2.object, fields)
}
\arguments{
\item{v2.object}{string ("/2/___"). Specifies an object for which a list of fields is returned.}

\item{fields}{vector of strings ("FIELD_1", ...). If specified, subsets the returned fields (accepts "all").}
}
\value{
A list of fields for the specified v2.object all collapsed with comma as separator (e.g. "id,text,attachments").
}
\description{
Getter for the fields of the objects available for request on the new Twitter API v2.
}
\author{
Victor Plesco
}
