% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_print_summary.R
\name{supp_print_summary}
\alias{supp_print_summary}
\title{supp_print_summary}
\usage{
supp_print_summary(process.data, twitter.chunk)
}
\arguments{
\item{token}{racademic's object. A bearer token associated with a user-created APP (requires a developer account).}

\item{v2.endpoint}{string \{"/2/___"\}.}
}
\value{

}
\description{

}
\author{
: Victor Plesco
}
