% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_validate_request.R
\name{supp_validate_request}
\alias{supp_validate_request}
\title{supp_validate_request}
\usage{
supp_validate_request(v2.endpoint, product_track, ...)
}
\arguments{
\item{v2.endpoint}{string \{"/2/___"\}. Specifies an endpoint whose query parameters are validated.}

\item{product_track}{string \{"academic" or "standard"\}. Specifies the type of developer account in possess.}

\item{...}{mix (e.g. max_results = 10, query = "btc -is:retweet", ect.). Query parameters to be validated.}
}
\value{
A list of query parameters.
}
\description{
Support method for validating the query parameters of a specified endpoint (given its query conditions).
}
\author{
: Victor Plesco
}
