% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_search_fullarchive.R
\name{v2_search_fullarchive}
\alias{v2_search_fullarchive}
\title{v2_search_fullarchive}
\usage{
v2_search_fullarchive(
  token,
  product_track,
  safe.dir = NULL,
  parse = FALSE,
  test = FALSE,
  query,
  start_time = NULL,
  end_time = NULL,
  max_results = 10,
  next_token = NULL,
  since_id = NULL,
  until_id = NULL,
  expansions = NULL,
  tweet.fields = NULL,
  user.fields = NULL,
  media.fields = NULL,
  place.fields = NULL,
  poll.fields = NULL
)
}
\arguments{
\item{token}{racademic's object. A bearer token associated with a user-created APP (requires a developer account).}

\item{product_track}{string \{"academic" or "standard"\}.}

\item{safe.dir}{string (e.g. "~/home/...").}

\item{parse}{boolean \{TRUE or FALSE \}.}

\item{query}{string (e.g. "btc -is:retweet").}

\item{start_time}{string \{%Y-%m-%dT%H:%M:%SZ\}.}

\item{end_time}{string \{%Y-%m-%dT%H:%M:%SZ\}.}

\item{max_results}{integer \[10, 500\].}

\item{next_token}{}

\item{since_id}{}

\item{until_id}{}

\item{expansions}{vector of strings (e.g. "expansion_\{1\}", ...). Specifies a subset of expansions to be returned (accepts "all"). 
To request all but some expansions, append "-" in front of each expansion in the input vector (e.g. "-expansion_\{1\}", ...).}

\item{___.fields}{vector of strings (e.g. "field_\{1\}", ...). Specifies a subset of fields to be returned (accepts "all"). To 
request all but some fields, append "-" in front of each field in the input vector (e.g. "-field_\{1\}", ...).}
}
\value{
A list or parsed format of tweets and relative data.
}
\description{

}
\author{
: Victor Plesco
}
